+++ lib/ruled/init.lua
@@ -1,1 +1,3 @@
-return {}
+return {
+    client = require("ruled.client");
+}
